introduce langdetect
