+++ pyproject.toml
@@ -18,6 +18,7 @@ twitter = "^1.18.0"
 gunicorn = "^20.0.4"
 python-dotenv = "^0.15.0"
 fcache = "^0.4.7"
+langdetect = "^1.0.8"
 
 [tool.poetry.dev-dependencies]
